requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
numpy-stl==2.12.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
six==1.15.0
